// Lakshya-One-main/frontend/src/lib/types/database.ts
export type Role = "PARENT" | "SCHOOL_ADMIN" | "ADMIN";

export type AdminAccessLevel = "READ_ONLY" | "READ_WRITE" | "FULL_ACCESS";